tools/pdf-extractor/requirements.txt
pymupdf
pytesseract
Pillow
sentence-transformers
faiss-cpu
langchain
